Remove Short and Word special case for unsigned 16 and 32
The use of Word and Short to designate data type is ambiguous, not only
they do not express the signed or unsigned nature of the type, the size
is also different depending on the architecture. Using UInt32 and UInt16
is also more coherent with the other unsigned integer types generated.

Signed-off-by: Fabien Chouteau <chouteau@adacore.com>