Add (modified) SVD files for nRF52832 and nRF52840.
Both SVD files have been manually modified to adjust the line lengths
in the `<licenseText>` element to ensure that the line length in the
generated code does not exceed 80 columns. This is necessary to avoid
style warnings when the generated code is used in the bb-runtimes repo.

The nrf52.svd file has additional modifications to add `<fields>`
elements to all `TASKS_*` and `EVENTS_*` register definitions so that
the generated register types are identical between the two devices.
This ensures that common user code accessing these registers can
compile using the generated code for either MCU.