Run macOS CI on working macOS version.
At present, macos-latest runs on macos-14 (which is aarch64, but that
doesn't affect this issue). The version of Xcode installed by default
on the Github runner is 15.0.1, which has the linking bug (assertion
failure in ld).

We could fix this by selecting a later version of Xcode (e.g.15.3),
but at some point this issue will be fixed and we can revert to
macos-latest. This seems the simplest solution, and will remain valid
until macos-12 is retired.

  * .github/workflows/main.yml (os): use macos-12.