Install SVDs.
When installing via Alire, include the schema and the SVD files.

Ideally, if a user provides just the name of the required SVD file,
svd2ada should search <executable-location>/../share/CMSIS-SVD/ for
the required file, but that'd be a bigger change; this way we at leat
have the SVDs on the user's disk.

  * alire.toml (version): updated to 0.1.1.
    (xmlada): updated to ">=22.0.0" (we're presently on 24.0.0).
  * svd2ada.gpr (Install): new.